requirements.txt
Django==1.10.3
django-crum==0.7.1
